pyproject.toml
[project]
name = "convert2md"
version = "0.1.0"
description = "Grab anything (URLs, GitHub, PDFs, YouTube) as clean Markdown for LLM context."
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "Rostand Kennezangue", email = "info@smarterservices.ai" }]
keywords = ["markdown", "llm", "pdf", "youtube", "chrome-extension", "cli"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
  "typer>=0.15",
  "rich>=13.9",
  "pydantic-settings>=2.14",
  "httpx[http2]>=0.28",
  "beautifulsoup4>=4.12",
  "lxml>=5",
  "pymupdf>=1.26",
  "youtube-transcript-api",
  "crawl4ai>=0.8",
]

[project.optional-dependencies]
gemini = ["google-genai>=1.0"]

[project.urls]
Homepage = "https://github.com/smarterservices-dot-ai/convert2md"
Issues = "https://github.com/smarterservices-dot-ai/convert2md/issues"
Contact = "mailto:info@smarterservices.ai"

[project.scripts]
convert2md = "convert2md.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["convert2md"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "C4"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Typer's Option defaults rely on mutable-default and call-in-default patterns.
"convert2md/cli.py" = ["B006", "B008"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
# crawl4ai ships no py.typed marker, so its DefaultMarkdownGenerator is Any.
# We subclass it intentionally for our adapter pass-through.
module = "convert2md.adapters"
disallow_subclassing_any = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "pytest-cov",
  "ruff",
  "mypy",
]
